Add Kerberos Attestor Agent
This change adds the Kerberos-Attestor Agent.  The agent works in
complement to the Server to attest new nodes.

With this change, we also clean up and add the required dependencies to
the glide files for both the server and agent attestor code.  It also
updates how code is organized between the server, agent, and common
modules.

Signed-off-by: Neel Shah <shahneel@vmware.com>